+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.4
+
+- Rename bin from `dartsc` to `scr`
+- Update help usage print format
+- Fix bug when pubspec.yaml is missing
+- Add line_length config
+
 ## 0.1.3
 
 - Update code formatting